Refactor as per comment

+++ src/main/java/io/github/jamsesso/jsonlogic/ast/JsonLogicNumber.java
@@ -14,7 +14,7 @@ public class JsonLogicNumber implements JsonLogicPrimitive<BigDecimal> {
 
   @Override
   public BigDecimal getValue() {
-    return (BigDecimal) value;
+    return value;
   }
 
   @Override
